+++ config/sus.rb
@@ -1,7 +1,10 @@
 # frozen_string_literal: true
 
 # Released under the MIT License.
-# Copyright, 2023-2024, by Samuel Williams.
+# Copyright, 2023-2025, by Samuel Williams.
+
+ENV["TRACES_BACKEND"] ||= "traces/backend/test"
+require "traces"
 
 require "covered/sus"
 include Covered::Sus